# HotelBooking.Client/API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HotelBooking.API",
    "version": "1.0"
  },
  "paths": {
    "/BookedRoom": {
      "get": {
        "tags": [
          "BookedRoom"
        ],
        "summary": "Получение информации о всех забронированных номерах.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookedRoomGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookedRoomGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookedRoomGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BookedRoom"
        ],
        "summary": "Добавление новой брони",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookedRoomDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BookedRoomDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BookedRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/BookedRoom/{id}": {
      "get": {
        "tags": [
          "BookedRoom"
        ],
        "summary": "Получение информации по id о комнате",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookedRoomGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookedRoomGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookedRoomGetDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "BookedRoom"
        ],
        "summary": "Изменение в брони",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookedRoomDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BookedRoomDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BookedRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "BookedRoom"
        ],
        "summary": "Удаление брони по Id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/BookedRoom/GetAllClientInHotel/{id}": {
      "get": {
        "tags": [
          "BookedRoom"
        ],
        "summary": "Возвращает всех клиентов в отеле",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/BookedRoom/GetAllFreeRooms/{city}": {
      "get": {
        "tags": [
          "BookedRoom"
        ],
        "summary": "Возвращает все свободные номера в выбранном городе",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Hotel": {
      "get": {
        "tags": [
          "Hotel"
        ],
        "summary": "Получить все отели",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HotelGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HotelGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HotelGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Hotel"
        ],
        "summary": "Добавить новый отель",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/HotelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Hotel/{id}": {
      "get": {
        "tags": [
          "Hotel"
        ],
        "summary": "Получить отель по Id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HotelGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelGetDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Hotel"
        ],
        "summary": "Изменияем старый отель",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/HotelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Hotel"
        ],
        "summary": "Удаляем отель",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/Room/{id}": {
      "get": {
        "tags": [
          "Room"
        ],
        "summary": "Получение информации о номере через id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RoomGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomGetDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Room"
        ],
        "summary": "Изменение данных о номере через id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Room"
        ],
        "summary": "Удаление номера по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BookedRoomDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "description": "Клиент",
            "format": "int32"
          },
          "roomId": {
            "type": "integer",
            "description": "Комната",
            "format": "int32"
          },
          "dateArrival": {
            "type": "string",
            "description": "Дата въезда",
            "nullable": true
          },
          "dateEvection": {
            "type": "string",
            "description": "Дата выселения",
            "nullable": true
          },
          "periodInDays": {
            "type": "integer",
            "description": "Период",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO класс для возвращения значений класса BookedRoom"
      },
      "BookedRoomGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "client": {
            "$ref": "#/components/schemas/ClientGetDto"
          },
          "room": {
            "$ref": "#/components/schemas/RoomGetDto"
          },
          "dateArrival": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "dateEvection": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "periodInDays": {
            "type": "integer",
            "description": "Период",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Client": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "passportData": {
            "$ref": "#/components/schemas/Passport"
          },
          "passportId": {
            "type": "integer",
            "format": "int32"
          },
          "birthOfDay": {
            "$ref": "#/components/schemas/DateOnly"
          }
        },
        "additionalProperties": false
      },
...
      "ClientGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "description": "ФИО",
            "nullable": true
          },
          "passportData": {
            "$ref": "#/components/schemas/PassportGetDto"
          },
          "birthOfDay": {
            "$ref": "#/components/schemas/DateOnly"
          }
        },
        "additionalProperties": false,
        "description": "DTO класс для возвращения значений класса Client"
      },
...
      "DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DayOfWeek": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "HotelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Адрес",
            "nullable": true
          },
          "city": {
            "type": "string",
            "description": "Город",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO класс для возвращения значений класса Hotel"
      },
      "HotelGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор отеля",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Адрес",
            "nullable": true
          },
          "city": {
            "type": "string",
            "description": "Город",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO класс для возвращения значений класса Hotel"
      },
      "Passport": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "series": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PassportGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор паспорта",
            "format": "int32"
          },
          "series": {
            "type": "integer",
            "description": "Серия",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "description": "Номер",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO класс для возвращения значений Passport"
      },
      "RoomDto": {
        "type": "object",
        "properties": {
          "typeId": {
            "type": "integer",
            "description": "Тип номера",
            "format": "int32"
          },
          "capacity": {
            "type": "integer",
            "description": "Вместимость",
            "format": "int32"
          },
          "cost": {
            "type": "number",
            "description": "Цена",
            "format": "double"
          },
          "hotelId": {
            "type": "integer",
            "description": "Id отеля",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO класс для возвращения значений Room"
      },
      "RoomGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/RoomGetDto"
          },
          "capacity": {
            "type": "integer",
            "description": "Вместимость",
            "format": "int32"
          },
          "cost": {
            "type": "number",
            "description": "Цена",
            "format": "double"
          },
          "hotelId": {
            "type": "integer",
            "description": "Id отеля",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO класс для возвращения значений Room"
      },
...
    }
  }
}